리팩터링 구조 변경
