--- app/[lang]/category/[category]/page.tsx
@@ -1,2 +1,0 @@
-import { restaurantService } from "@/shared/api/restaurantService";
-// ... 다른 imports
