+++ app/[lang]/restaurant/[id]/page.tsx
@@ -1,20 +1,17 @@
 import { restaurantApi } from "@/shared/api/restaurants";
 import { RestaurantDetail } from "@/widgets/restaurant/ui/RestaurantDetail";
 import { notFound } from "next/navigation";
-import { DetailHeader } from "@/app/components/DetailHeader";
+import { DetailHeader } from "@/entities/restaurant/ui/DetailHeader";
 
 interface RestaurantDetailPageProps {
-  params: Promise<{ id: string }> | undefined;
+  params: {
+    id: string;
+  };
 }
 
 export default async function RestaurantDetailPage({
-  params,
+  params: { id },
 }: RestaurantDetailPageProps) {
-  if (!params) {
-    notFound();
-  }
-
-  const { id } = await params;
   const restaurant = await restaurantApi.getRestaurantDetail(id);
 
   if (!restaurant) {
@@ -23,7 +20,10 @@ export default async function RestaurantDetailPage({
 
   return (
     <div className="fixed inset-0 bg-orange-100 z-50 overflow-y-auto">
-      <DetailHeader title={restaurant.name} rating={restaurant.rating} />
+      <DetailHeader
+        title={restaurant.name}
+        rating={restaurant.rating ?? undefined}
+      />
       <div className="mx-auto w-full max-w-[480px] pt-4">
         <RestaurantDetail restaurant={restaurant} />
       </div>
