+++ app/i18n/dictionaries/zh.json
@@ -1,0 +1,11 @@
+{
+  "recommended": "推薦",
+  "restaurantList": "餐廳列表",
+  "themeNotFoundTitle": "主題不存在",
+  "themeNotFoundDescription": "找不到請求的主題。",
+  "chooseAnotherTheme": "請選擇其他主題。",
+  "home": "返回首頁",
+  "theme": "主題",
+  "noRestaurantsFoundTitle": "沒有餐廳",
+  "noRestaurantsFoundDescription": "該主題下尚未添加任何餐廳。"
+}
