+++ entities/restaurant/ui/DetailHeader.tsx
